fixes to populate environment in config
