+++ duradmin/src/main/java/org/duracloud/duradmin/control/LoginController.java
@@ -7,8 +7,9 @@
  */
 package org.duracloud.duradmin.control;
 
-import org.duracloud.common.util.DuracloudConfig;
+import org.duracloud.common.util.DuracloudConfigBean;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
@@ -20,10 +21,12 @@ import org.springframework.web.servlet.ModelAndView;
 @Controller
 public class LoginController {
 
+    @Autowired
+    private DuracloudConfigBean duracloudConfigBean;
+
     @RequestMapping("/login")
     public ModelAndView presentLogin() throws Exception {
-        DuracloudConfig duracloudConfig = new DuracloudConfig();
-        String amaUrl = duracloudConfig.getAmaUrl();
+        String amaUrl = duracloudConfigBean.getAmaUrl();
 
         ModelAndView login = new ModelAndView("login");
         login.addObject("amaUrl", amaUrl);
